webapp/FastAPI/requirements-mac.txt
accelerate==1.6.0
annotated-types==0.7.0
anyio==4.8.0
av==14.3.0
bcrypt==4.3.0
bitsandbytes==0.47.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.7.2
fastapi==0.115.11
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.3.1
h11==0.14.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
motor==3.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
open_clip_torch==2.31.0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.0
transformers-stream-generator==0.0.5
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
language_tool_python==2.9.3
deep-translator==1.11.4
sentence-transformers==4.1.0
pandas==2.2.3
hf-xet==1.1.2